refac: move the core webauthn code to ~packages/ folder
