+++ background/messages/mCreateWindow.ts
@@ -9,7 +9,7 @@ import {
 import type {
   WebauthnCreation,
   WebauthnRequest
-} from "~packages/account/PasskeyAccount/passkeyOwnerWebauthn/webauthn/typing"
+} from "~packages/webauthn/typing"
 
 export type RequestBody = {
   creation: WebauthnCreation
