+++ contents/messages.ts
@@ -30,7 +30,7 @@ import {
 import type {
   WebauthnCreation,
   WebauthnRequest
-} from "~packages/account/PasskeyAccount/passkeyOwnerWebauthn/webauthn/typing"
+} from "~packages/webauthn/typing"
 import type { UrlB64String } from "~typing"
 
 const ethersAbi = AbiCoder.defaultAbiCoder()
