+++ packages/webauthn/typing.ts
